+++ http/client/evaluator.go
