+++ http/client/evaluator_test.go
