+++ http/client/rulesets_test.go
